fork

+++ src/components/Header.jsx
@@ -1,61 +1,61 @@
+/* eslint-disable @next/next/no-img-element */
+"use client"
 import React, { useState, useEffect } from "react";
-import "../styles/components/header.css";
 import { MdMenu, MdClear } from "react-icons/md";
-import LogoWhite from "../assets/img/logo-white.svg";
-import { Outlet, Link } from "react-router-dom";
-import Pdf from "../assets/dopot.pdf";
 import { SocialIcon } from "react-social-icons";
-import IconDown from "../assets/img/arr-menu.svg";
-import { useTranslation } from "react-i18next";
+import { useTranslation } from "../i18n/client";
 import { FaTimes } from 'react-icons/fa';
-
-//import GetAccount from "../utils/ethersUtils.js";
 import { getAddr } from "../utils/firebase/retriveInfo";
 import { ethers } from "ethers";
+import Link from 'next/link';
 
 async function isWalletConnected() {
-  const provider = new ethers.providers.Web3Provider(window.ethereum);
-  return provider && (await provider.send("eth_accounts", [])).length > 0;
+  if (typeof window !== "undefined" && window.ethereum) {
+    const provider = new ethers.providers.Web3Provider(window.ethereum);
+    return provider && (await provider.send("eth_accounts", [])).length > 0;
+  }
 }
 
-function deleteCookiesAndReload() {
+function deleteDatabase(dbName) {
+  return new Promise((resolve, reject) => {
+    const deleteRequest = indexedDB.deleteDatabase(dbName);
+    deleteRequest.onsuccess = () => resolve(`Deleted ${dbName} successfully`);
+    deleteRequest.onerror = (event) => new Error(`Error deleting ${dbName}: ${event.target.errorCode}`);
+  });
+}
+
+const DeleteCookiesAndReload = () => {
   localStorage.clear();
   // Delete all cookies
-  document.cookie.split(";").forEach(function(c) {
-    document.cookie = c
+  if(typeof window !== "undefined"){
+    document.cookie.split(";").forEach(function(c) {
+      document.cookie = c
       .replace(/^ +/, "")
       .replace(/=.*/, "=;expires=" + new Date().toUTCString() + ";path=/");
-  });
-  // Clear IndexedDB
-  const indexedDB = window.indexedDB || window.mozIndexedDB || window.webkitIndexedDB || window.msIndexedDB;
-  function deleteDatabase(dbName) {
-    return new Promise((resolve, reject) => {
-      const deleteRequest = indexedDB.deleteDatabase(dbName);
-      deleteRequest.onsuccess = () => resolve(`Deleted ${dbName} successfully`);
-      deleteRequest.onerror = (event) => reject(`Error deleting ${dbName}: ${event.target.errorCode}`);
     });
-  }
+    // Clear IndexedDB
+    const indexedDB = window.indexedDB || window.mozIndexedDB || window.webkitIndexedDB || window.msIndexedDB;
     indexedDB.databases().then((dbs) => {
-    const deletePromises = dbs.map((db) => deleteDatabase(db.name));
-    return Promise.all(deletePromises);
-  }).then((results) => {
-    console.log(results);
-    window.location.reload();
-  }).catch((error) => {
-    console.error(error);
-    window.location.reload();
-  });
+      const deletePromises = dbs.map((db) => deleteDatabase(db.name));
+      return Promise.all(deletePromises);
+    }).then((results) => {
+      console.log(results);
+      window.location.reload();
+    }).catch((error) => {
+      console.error(error);
+    });
+    }
 }
 
 const Header = (props) => {
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
   const [isHeaderOpen, setIsHeaderOpen] = useState(false);
-  const [walletText, setwalletText] = useState("Connect Wallet");
+  const [walletText, setWalletText] = useState("Connect Wallet");
   const [walletState, setWalletState] = useState(false);
   useEffect(() => {
     (async () => {
       setWalletState(await isWalletConnected());
-      getAddr(setwalletText, true, t);
+      if(walletState) getAddr(setWalletText, true, t);
     })();
   });
   return (
@@ -64,65 +64,46 @@ const Header = (props) => {
         {/* Header for PC */}
         <div className="header-content">
           <div className="header-left">
-            <a href="/#/">
-              <img src={LogoWhite} alt="LogoWhite" />
-            </a>
+            <Link href="/">
+              <img src={"/assets/img/logo-white.svg"} alt="Dopot" />
+            </Link>
           </div>
           <div className="header-right">
-            <a href="/#/">Home</a>
-            <a href="/#/FaqIta">Tutorials</a>
-            <a href="/#/ambassador">Ambassador</a>
-            <a href={"/#/dopottoken"}>Dpt Token</a>
+            <Link href="/">Home</Link>
+            <Link href="/FaqIta">Tutorials</Link>
+            <Link href="/Ambassador">Ambassador</Link>
+            <Link href="/DopotToken">Dpt Token</Link>
 
             <div style={{ marginRight: "1.5rem" }} className="dropdown_menu">
               <button className="dropbtn">
                 Community{" "}
                 <span>
-                  <img className="arrow-menu-dr" src={IconDown} alt="" />{" "}
+                  <img className="arrow-menu-dr" src={"/assets/img/arr-menu.svg"} alt="" />{" "}
                 </span>
               </button>
 
               <div className="dropdown-content-menu">
-                <a>
+                <button>
                   <SocialIcon
                     fgColor="white"
                     url="https://www.instagram.com/dopotfi"
                   />
-                </a>
-                {/* <a>
-                  <SocialIcon
-                    fgColor="white"
-                    url="https://www.facebook.com/dopot.fi"
-                  />
-                </a> */}
-                <a>
+                </button>
+                <button>
                   <SocialIcon
                     fgColor="white"
                     url="https://twitter.com/Dopot_fi"
                   />
-                </a>
-                {/* <a>
-                  <SocialIcon
-                    fgColor="white"
-                    bgColor="#EE1D52"
-                    url="https://www.tiktok.com/@dopotfi"
-                  />
-                </a> 
-                <a>
-                  <SocialIcon
-                    fgColor="white"
-                    url="https://reddit.com/r/dopotfi/s/qWC4ui9Fl9"
-                  />
-                </a>*/}
-                <a>
+                </button>
+                <button>
                   <SocialIcon
                     fgColor="white"
                     url="https://discord.gg/j8xxZFsyvd"
                   />
-                </a>
-                <a>
+                </button>
+                <button>
                   <SocialIcon fgColor="white" url="https://t.me/dopotfi" />
-                </a>
+                </button>
               </div>
             </div>
             <div style={{ marginRight: "3.5rem" }} className="dropdown_menu">
@@ -132,115 +113,96 @@ const Header = (props) => {
                   <img
                     style={{ width: "7%" }}
                     className="arrow-menu-dr"
-                    src={IconDown}
+                    src={"/assets/img/arr-menu.svg"}
                     alt=""
                   />
                 </span>
               </button>
               <div className="dropdown-content-menu">
-                <a href={Pdf}>Whitepaper</a>
-                <a href="https://dopot.gitbook.io/dopot/">Gitbook</a>
-                <a href="https://github.com/solidproof/projects/blob/main/2024/Dopot/SmartContract_Audit_Solidproof_DopotFi.pdf">Audit</a>
+                <Link href="/assets/dopot.pdf">Whitepaper</Link>
+                <Link href="https://dopot.gitbook.io/dopot/">Gitbook</Link>
+                <Link href="https://github.com/solidproof/projects/blob/main/2024/Dopot/SmartContract_Audit_Solidproof_DopotFi.pdf">Audit</Link>
               </div>
             </div>
             {walletState && (
-              <a href={"/#/profile"}>
+              <Link href="/Profile">
                 <button className="grd-btn dopot-btn-sm">Account</button>
-              </a>
+              </Link>
             )}
 
             <button
               className="purple-border-btn dopot-btn-sm"
-              onClick={() => getAddr(setwalletText, false, t)}
+              onClick={() => getAddr(setWalletText, false, t)}
             >
               {walletText}
             </button>
-            <a href="/#/" onClick={deleteCookiesAndReload}>
+            <Link href="/" onClick={DeleteCookiesAndReload}>
               <FaTimes />
-            </a>
+            </Link>
           </div>
         </div>
         {/* Header for Mobile Devices */}
         <div className="header-mob">
-          <a href="#">
-            <img src={LogoWhite} alt="LogoWhite" />
-          </a>
+          <Link href="/">
+            <img src={"/assets/img/logo-white.svg"} alt="Dopot" />
+          </Link>
 
           {isHeaderOpen ? (
-            <div
+            <button
               className="header-icon"
               onClick={() => {
                 setIsHeaderOpen(false);
               }}
             >
               <MdClear />
-            </div>
+            </button>
           ) : (
-            <div
+            <button
               className="header-icon"
               onClick={() => {
                 setIsHeaderOpen(true);
               }}
             >
               <MdMenu />
-            </div>
+            </button>
           )}
 
           {isHeaderOpen ? (
             <div className="header-mob-box">
-              <a href="#">Home</a>
-              <a href="/#/FaqIta">Tutorials</a>
-              <a href={"/#/dopottoken"}>DPT Token</a>
-              <a href="/#/ambassador">Ambassador</a>
+              <Link href="/">Home</Link>
+              <Link href="/FaqIta">Tutorials</Link>
+              <Link href="/DopotToken">DPT Token</Link>
+              <Link href="/Ambassador">Ambassador</Link>
               <div style={{ marginRight: "1.5rem" }} className="dropdown_menu">
                 <button className="dropbtn">
                   Community{" "}
                   <span>
-                    <img className="arrow-menu-dr" src={IconDown} alt="" />{" "}
+                    <img className="arrow-menu-dr" src={"/assets/img/arr-menu.svg"} alt="" />{" "}
                   </span>
                 </button>
 
                 <div className="dropdown-content-menu">
-                  <a>
+                  <button>
                     <SocialIcon
                       fgColor="white"
                       url="https://www.instagram.com/dopotfi"
                     />
-                  </a>
-                  {/* <a>
-                    <SocialIcon
-                      fgColor="white"
-                      url="https://www.facebook.com/dopot.fi"
-                    />
-                  </a> */}
-                  <a>
+                  </button>
+                  <button>
                     <SocialIcon
                       fgColor="white"
                       url="https://twitter.com/Dopot_fi"
                     />
-                  </a>
-                  {/* <a>
-                     <SocialIcon
-                      fgColor="white"
-                      bgColor="#EE1D52"
-                      url="https://www.tiktok.com/@dopotfi"
-                    /> 
-                  </a>
-                  <a>
-                    <SocialIcon
-                      fgColor="white"
-                      url="https://reddit.com/r/dopotfi/s/qWC4ui9Fl9"
-                    />
-                  </a>*/}
-                  <a>
+                  </button>
+                  <button>
                     <SocialIcon
                       fgColor="white"
                       url="https://discord.gg/j8xxZFsyvd"
                     />
-                  </a>
-                  <a>
+                  </button>
+                  <button>
                     <SocialIcon fgColor="white" url="https://t.me/dopotfi" />
-                  </a>
+                  </button>
                 </div>
               </div>
               <div style={{ marginRight: "3.5rem" }} className="dropdown_menu">
@@ -250,15 +212,15 @@ const Header = (props) => {
                     <img
                       style={{ width: "7%" }}
                       className="arrow-menu-dr"
-                      src={IconDown}
+                      src={"/assets/img/arr-menu.svg"}
                       alt=""
                     />
                   </span>
                 </button>
                 <div className="dropdown-content-menu">
-                  <a href={Pdf}>Whitepaper</a>
-                  <a href="https://dopot.gitbook.io/dopot/">Gitbook</a>
-                  <a href="https://github.com/solidproof/projects/blob/main/2024/Dopot/SmartContract_Audit_Solidproof_DopotFi.pdf">Audit</a>
+                  <Link href="/assets/dopot.pdf">Whitepaper</Link>
+                  <Link href="https://dopot.gitbook.io/dopot/">Gitbook</Link>
+                  <Link href="https://github.com/solidproof/projects/blob/main/2024/Dopot/SmartContract_Audit_Solidproof_DopotFi.pdf">Audit</Link>
                 </div>
               </div>
